events_service/pyproject.toml
[project]
name = "events-service"
version = "0.1.0"
description = ""
authors = [
    {name = "Artem Zeinalov",email = "zeynchan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [
    {include = "app"},
    {include = "core"}
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
